backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic-settings==2.7.0
sqlalchemy==2.0.36
alembic==1.14.0
apscheduler==3.10.4
